API_ENDPOINT/requirements.txt
aniso8601==6.0.0
arrow==0.14.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.7
flask-redis==0.4.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
flask-swagger-ui==3.20.9
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pip-autoremove==0.9.1
pkg-resources==0.0.0
psycopg2==2.8.2
python-dateutil==2.8.0
python-dotenv==0.10.2
pytz==2019.1
PyYAML==5.1
redis==3.2.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
